Improve base tests and readme

+++ tests/test.bats
@@ -1,31 +1,36 @@
 setup() {
   export DIR="$( cd "$( dirname "$BATS_TEST_FILENAME" )" >/dev/null 2>&1 && pwd )/.."
-  export TESTDIR=$(mktemp -d -t testmemcached-XXXXXXXXXX)
-  export PROJNAME=testmemcached
+  export TESTDIR=~/tmp/testelasticsearch
+  mkdir -p $TESTDIR
+  export PROJNAME=test-addon-template
   export DDEV_NON_INTERACTIVE=true
   ddev delete -Oy ${PROJNAME} || true
   cd "${TESTDIR}"
-  ddev config --project-name=${PROJNAME} --project-type=drupal9 --docroot=web --create-docroot
+  ddev config --project-name=${PROJNAME}
   ddev start
 }
 
 teardown() {
-  cd ${TESTDIR}
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
   ddev delete -Oy ${DDEV_SITENAME}
-  rm -rf ${TESTDIR}
+  [ "${TESTDIR}" != "" ] && rm -rf ${TESTDIR}
 }
 
-@test "basic installation" {
+@test "install from directory" {
   cd ${TESTDIR}
+  echo "# ddev get ${DIR} with project ${PROJNAME} in ${TESTDIR} ($(pwd))" >&3
   ddev get ${DIR}
   ddev restart
-#  v=$(ddev exec 'printf "version\nquit\nquit\n" | nc memcached 11211')
-#  [[ "${v}" = VERSION* ]]
-#  res=$(ddev exec 'printf "list-tubes\nquit\n" | nc -C beanstalkd 11300')
-#  [[ "${res}" = OK* ]]
-#  status=$(ddev exec 'drush sapi-sl --format=json | jq -r .default_solr_server.status')
-#  [ "${status}" = "enabled" ]
-#  sleep 10 # After a restart, the solr server may not be ready yet.
-#  ddev drush search-api-solr:reload default_solr_server
+  # Do something here to verify functioning extra service
+  # For extra credit, use a real CMS with actual config.
+  # ddev exec "curl -s elasticsearch:9200" | grep "${PROJNAME}-elasticsearch"
 
+}
+
+@test "install from release" {
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
+  echo "# ddev get drud/ddev-addon-template with project ${PROJNAME} in ${TESTDIR} ($(pwd))" >&3
+  ddev get drud/ddev-addon-template
+  ddev restart
+  # ddev exec "curl -s elasticsearch:9200" | grep "${PROJNAME}-elasticsearch"
 }
